requirements.txt
numpy>=1.26
scipy>=1.11
matplotlib>=3.8
pandas>=2.2
pyserial>=3.5
PyYAML>=6.0
